// apps/api/src/content/content.types.ts
export enum ContentType {
  OFFICIAL_NEWS = 'OFFICIAL_NEWS',
  OFFICIAL_WARNING = 'OFFICIAL_WARNING',
  OFFICIAL_EVENT = 'OFFICIAL_EVENT',
  USER_POST = 'USER_POST',
  MARKETPLACE_LISTING = 'MARKETPLACE_LISTING',
  HELP_REQUEST = 'HELP_REQUEST',
  HELP_OFFER = 'HELP_OFFER',
  MOVING_CLEARANCE = 'MOVING_CLEARANCE',
  CAFE_MEETUP = 'CAFE_MEETUP',
  KIDS_MEETUP = 'KIDS_MEETUP',
  APARTMENT_SEARCH = 'APARTMENT_SEARCH',
  LOST_FOUND = 'LOST_FOUND',
  RIDE_SHARING = 'RIDE_SHARING',
  JOBS_LOCAL = 'JOBS_LOCAL',
  VOLUNTEERING = 'VOLUNTEERING',
  GIVEAWAY = 'GIVEAWAY',
  SKILL_EXCHANGE = 'SKILL_EXCHANGE',
}

export enum Category {
  MARKETPLACE = 'MARKETPLACE',
  HELP_REQUEST = 'HELP_REQUEST',
  HELP_OFFER = 'HELP_OFFER',
  MOVING_CLEARANCE = 'MOVING_CLEARANCE',
  CAFE_MEETUP = 'CAFE_MEETUP',
  KIDS_MEETUP = 'KIDS_MEETUP',
  APARTMENT_SEARCH = 'APARTMENT_SEARCH',
  LOST_FOUND = 'LOST_FOUND',
  RIDE_SHARING = 'RIDE_SHARING',
  JOBS_LOCAL = 'JOBS_LOCAL',
  VOLUNTEERING = 'VOLUNTEERING',
  GIVEAWAY = 'GIVEAWAY',
  SKILL_EXCHANGE = 'SKILL_EXCHANGE',
}